pyproject.toml
[tool.poetry]
name = "tlo"
version = "0.0.1"
description = "This project includes an application that allows users to configure traffic objects (e.g., intersections) equipped with traffic lights, using their own parameters such as signal phase length, capacity, number of lanes, and traffic flow. The application then provides an optimized traffic light scheme that minimizes traffic congestion."
authors = ["Andrew <freetown113@gmail.com>"]
readme = "README.md"
packages = [
  { include = "tlo", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^1.13.0"
gymnasium = "^0.29.1"
stable-baselines3 = "^2.1.0"
sumo-rl = "^1.4.3"
django = "^4.2.5"
celery = "^5.3.4"
redis = "^5.0.1"
flower = "^2.0.1"
dvc = "^3.23.0"
python-box = "^7.1.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"
bandit = "^1.7.5"
flake8 = "^6.1.0"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pre-commit = "^3.4.0"
monkeytype = "^23.3.0"
types-pyyaml = "^6.0.12.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
